+++ data/spatial/population.py
@@ -34,7 +34,7 @@ def execute(context):
     requested_iris = set(df_codes["iris_id"].unique())
     merged_iris = set(df_population["iris_id"].unique())
 
-    if len(requested_iris) != len(merged_iris):
+    if requested_iris != merged_iris:
         raise RuntimeError("Some IRIS are missing: %s" % (requested_iris - merged_iris,))
 
     return df_population[["region_id", "departement_id", "commune_id", "iris_id", "population"]]